// apps/feeds/frontend/src/app/graphql/types.ts
export type Maybe<T> = T | null
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
  LocalDateTime: any
}

export type Article = {
  __typename?: 'Article'
  id: Scalars['ID']
  link: Scalars['String']
  title: Scalars['String']
  author: Scalars['String']
  content: Scalars['String']
  publishedDate: Scalars['LocalDateTime']
  feed: Feed
  unread: Scalars['Boolean']
}

export type ArticleFilterCriteria = {
  feedIds?: Maybe<Array<Scalars['ID']>>
  offset?: Maybe<Scalars['Int']>
  limit?: Maybe<Scalars['Int']>
}

export type Feed = {
  __typename?: 'Feed'
  id: Scalars['ID']
  uri: Scalars['String']
  title: Scalars['String']
  icon?: Maybe<Scalars['String']>
  articles: Array<Article>
  group?: Maybe<Group>
  unreadCount?: Maybe<Scalars['Int']>
  totalCount?: Maybe<Scalars['Int']>
  hasFetchError: Scalars['Boolean']
}

export type FeedArticlesArgs = {
  unread?: Maybe<Scalars['Boolean']>
  offset?: Maybe<Scalars['Int']>
  limit?: Maybe<Scalars['Int']>
}

export type FeedInput = {
  uri: Scalars['String']
  title?: Maybe<Scalars['String']>
}

export type Group = {
  __typename?: 'Group'
  id: Scalars['ID']
  name: Scalars['String']
  feeds: Array<Feed>
}
// ...
export type Query = {
  __typename?: 'Query'
  articles: Array<Article>
  feeds: Array<Feed>
  groups: Array<Group>
  authorizationUri: Scalars['String']
  login?: Maybe<Token>
  me?: Maybe<User>
}

export type QueryArticlesArgs = {
  filter: ArticleFilterCriteria
}

export type QueryLoginArgs = {
  code: Scalars['String']
}

export type Subscription = {
  __typename?: 'Subscription'
  feed?: Maybe<Feed>
}

export type Token = {
  __typename?: 'Token'
  access_token: Scalars['String']
  refresh_token: Scalars['String']
  scope: Scalars['String']
  token_type: Scalars['String']
}

export type User = {
  __typename?: 'User'
  username: Scalars['String']
  name?: Maybe<Scalars['String']>
  avatar_url?: Maybe<Scalars['String']>
  email?: Maybe<Scalars['String']>
}

export type ArticleFieldsFragment = { __typename: 'Article' } & Pick<
  Article,
  'id' | 'title' | 'link' | 'author' | 'content' | 'publishedDate' | 'unread'
> & { feed: { __typename: 'Feed' } & Pick<Feed, 'id'> }
// ...
export type ArticlesQueryVariables = {
  filter: ArticleFilterCriteria
}

export type ArticlesQuery = { __typename?: 'Query' } & {
  articles: Array<{ __typename?: 'Article' } & ArticleFieldsFragment>
}